+++ library/admin.php
@@ -76,7 +76,7 @@ Customize it, and I'll advertise it.
 /*
 I don't really recommend editing the admin too much
 as things may get wonky if WordPress updates. Here
-are a few funtions which you can choose to use if
+are a few functions which you can choose to use if
 you like.
 */